+++ Changelog.rst
@@ -3,6 +3,23 @@ Version NEXTVERSION
 
 **2026-??-??**
 
+* Support for HEALPix grids
+  (https://github.com/NCAS-CMS/cf-python/issues/909)
+* New HEALPix methods: `cf.Field.healpix_info`,
+  `cf.Field.healpix_decrease_refinement_level`,
+  `cf.Field.healpix_increase_refinement_level`,
+  `cf.Field.healpix_change_indexing_scheme`,
+  `cf.Field.healpix_to_ugrid`, `cf.Domain.create_healpix`
+  (https://github.com/NCAS-CMS/cf-python/issues/909)
+* New method: `cf.Field.create_latlon_coordinates`
+  (https://github.com/NCAS-CMS/cf-python/issues/909)
+* New method: `cf.Data.coarsen`
+  (https://github.com/NCAS-CMS/cf-python/issues/909)
+* New functions: `cf.locate`, `cf.healpix_max_refinement_level`,
+  `cf.healpix_indexing_schemes`
+  (https://github.com/NCAS-CMS/cf-python/issues/909)
+* New keyword to `cf.read`: ``filesystem``
+  (https://github.com/NCAS-CMS/cf-python/issues/931)
 * New keyword parameter to `cf.Data.compute`: ``persist``
   (https://github.com/NCAS-CMS/cf-python/issues/929)
 * New function to control the persistence of computed data:
@@ -22,7 +39,11 @@ Version NEXTVERSION
   (https://github.com/NCAS-CMS/cfdm/issues/391)
 * Fix for subspacing with cyclic `cf.wi` and `cf.wo` arguments
   (https://github.com/NCAS-CMS/cf-python/issues/887)
-* Changed dependency: ``cfdm>=1.13.1.0, <1.13.2.0``
+* New optional dependency: ``healpix>=2025.1``
+* New dependency: ``pyfive>=1.1.1``
+* Changed dependency: ``cfdm>=1.13.?.?, <1.13.?.0``
+
+----
 
 Version 3.19.0
 --------------
@@ -79,8 +100,8 @@ Version 3.18.1
 * Allow multiple conditions for the same axis in `cf.Field.subspace`
   and `cf.Field.indices`
   (https://github.com/NCAS-CMS/cf-python/issues/881)
-* Fix bug in `cf.Field.collapse` that causes a ``ValueError`` to be raised
-  for missing external cell measures data
+* Fix bug in `cf.Field.collapse` that causes a ``ValueError`` to be
+  raised for missing external cell measures data
   (https://github.com/NCAS-CMS/cf-python/issues/885)
 * New dependency: ``distributed>=2025.5.1``
 * Changed dependency: ``cfdm>=1.12.3.0, <1.12.4.0``